+++ R/checkArgumentsPDTMT.R
@@ -9,15 +9,16 @@
     templateRmd, outputDir, outputBaseName, reportTitle, reportAuthor,
     forceOverwrite, experimentInfo, species, pdOutputFolder, pdResultName,
     inputLevel, pdAnalysisFile, idCol, labelCol, geneIdCol, proteinIdCol,
-    stringIdCol, iColPattern, sampleAnnot, includeOnlySamples, excludeSamples,
+    stringIdCol, modificationsCol, excludeUnmodifiedPeptides,
+    keepModifications, iColPattern, sampleAnnot, includeOnlySamples, excludeSamples,
     minScore, minDeltaScore, minPeptides, minPSMs, masterProteinsOnly,
     imputeMethod, mergeGroups,
     comparisons, ctrlGroup, allPairwiseComparisons, singleFit,
     subtractBaseline, baselineGroup, normMethod, spikeFeatures, stattest,
     minNbrValidValues, minlFC, samSignificance, nperm, volcanoAdjPvalThr,
     volcanoLog2FCThr, volcanoMaxFeatures, volcanoS0, volcanoFeaturesToLabel,
-    addInteractiveVolcanos, interactiveDisplayColumns, complexFDRThr,
-    maxNbrComplexesToPlot, seed,
+    addInteractiveVolcanos, interactiveDisplayColumns, interactiveGroupColumn,
+    complexFDRThr, maxNbrComplexesToPlot, seed,
     includeFeatureCollections, minSizeToKeepSet, customComplexes,
     complexSpecies, complexDbPath, stringVersion, stringDir, linkTableColumns,
     customYml, doRender, generateQCPlot
@@ -141,6 +142,11 @@
     } else if (inputLevel == "PeptideGroups") {
         .assertScalar(x = minDeltaScore, type = "numeric", allowNULL = TRUE)
         .assertScalar(x = minPSMs, type = "numeric", allowNULL = TRUE)
+
+        .assertScalar(x = modificationsCol, type = "character",
+                      allowNULL = TRUE)
+        .assertScalar(x = excludeUnmodifiedPeptides, type = "logical")
+        .assertScalar(x = keepModifications, type = "character", allowNULL = TRUE)
     }
 
     ## Method choices
@@ -175,6 +181,7 @@
     .assertScalar(x = allPairwiseComparisons, type = "logical")
     .assertScalar(x = addInteractiveVolcanos, type = "logical")
     .assertVector(x = interactiveDisplayColumns, type = "character", allowNULL = TRUE)
+    .assertScalar(x = interactiveGroupColumn, type = "character", allowNULL = TRUE)
     .assertScalar(x = singleFit, type = "logical")
     .assertScalar(x = subtractBaseline, type = "logical")
     .assertScalar(x = baselineGroup, type = "character")