+++ R/checkArgumentsPDTMTptm.R
@@ -9,13 +9,12 @@
         templateRmd, outputDir, outputBaseName, reportTitle, reportAuthor, forceOverwrite,
         experimentInfo, species, sceProteins, scePeptides, assayForTests,
         assayImputation, idCol, labelCol, proteinIdColProteins,
-        proteinIdColPeptides, modificationsCol, excludeUnmodifiedPeptides,
-        keepModifications, comparisons, ctrlGroup, allPairwiseComparisons,
+        proteinIdColPeptides, comparisons, ctrlGroup, allPairwiseComparisons,
         singleFit, subtractBaseline, baselineGroup,
         testType, minNbrValidValues, minlFC, volcanoAdjPvalThr, volcanoLog2FCThr,
         volcanoMaxFeatures, volcanoFeaturesToLabel,
-        addInteractiveVolcanos, interactiveDisplayColumns, seed,
-        linkTableColumns, customYml, doRender
+        addInteractiveVolcanos, interactiveDisplayColumns,
+        interactiveGroupColumn, seed, linkTableColumns, customYml, doRender
 ) {
     ## templateRmd
     .assertScalar(x = templateRmd, type = "character")
@@ -77,10 +76,6 @@
 
     .assertVector(x = linkTableColumns, type = "character", allowNULL = TRUE)
 
-    .assertScalar(x = modificationsCol, type = "character")
-    .assertScalar(x = excludeUnmodifiedPeptides, type = "logical")
-    .assertScalar(x = keepModifications, type = "character", allowNULL = TRUE)
-
     ## Method choices
     .assertScalar(x = testType, type = "character",
                   validValues = c("interaction", "welch"))
@@ -97,6 +92,7 @@
     .assertScalar(x = allPairwiseComparisons, type = "logical")
     .assertScalar(x = addInteractiveVolcanos, type = "logical")
     .assertVector(x = interactiveDisplayColumns, type = "character", allowNULL = TRUE)
+    .assertScalar(x = interactiveGroupColumn, type = "character", allowNULL = TRUE)
     .assertScalar(x = singleFit, type = "logical")
     .assertScalar(x = subtractBaseline, type = "logical")
     .assertScalar(x = baselineGroup, type = "character")