+++ tests/testthat/test-checkArgumentsPDTMTptm.R
@@ -27,9 +27,6 @@ test_that("argument checking for PD-TMT PTM works", {
                                            splitSeparator = ";", joinSeparator = "."),
         proteinIdColProteins = "einprotProtein",
         proteinIdColPeptides = "einprotProtein",
-        modificationsCol = "Modifications.in.Master.Proteins",
-        excludeUnmodifiedPeptides = FALSE,
-        keepModifications = NULL,
         comparisons = list(),
         ctrlGroup = "",
         allPairwiseComparisons = TRUE,
@@ -45,6 +42,7 @@ test_that("argument checking for PD-TMT PTM works", {
         volcanoFeaturesToLabel = c(""),
         addInteractiveVolcanos = FALSE,
         interactiveDisplayColumns = NULL,
+        interactiveGroupColumn = NULL,
         seed = 123,
         linkTableColumns = c(),
         customYml = NULL,
@@ -199,30 +197,6 @@ test_that("argument checking for PD-TMT PTM works", {
     expect_error(do.call(.checkArgumentsPDTMTptm, args),
                  "length(formals(proteinIdColPeptides)) == 1 is not TRUE", fixed = TRUE)
 
-    ## modificationsCol
-    args <- args0
-    args$modificationsCol <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'modificationsCol' must be of class 'character'")
-    args$modificationsCol <- c("Modifications", "Modifications.in.Master.Proteins")
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'modificationsCol' must have length 1")
-
-    ## excludeUnmodifiedPeptides
-    args <- args0
-    args$excludeUnmodifiedPeptides <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'excludeUnmodifiedPeptides' must be of class 'logical'")
-    args$excludeUnmodifiedPeptides <- c(TRUE, FALSE)
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'excludeUnmodifiedPeptides' must have length 1")
-
-    ## keepModifications
-    args <- args0
-    args$keepModifications <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "'keepModifications' must be of class 'character'")
-
     ## comparisons
     args <- args0
     args$comparisons <- 1
@@ -374,6 +348,15 @@ test_that("argument checking for PD-TMT PTM works", {
     args$interactiveDisplayColumns <- 1
     expect_error(do.call(.checkArgumentsPDTMTptm, args),
                  "'interactiveDisplayColumns' must be of class 'character'")
+
+    # interactiveGroupColumn
+    args <- args0
+    args$interactiveGroupColumn <- 1
+    expect_error(do.call(.checkArgumentsPDTMTptm, args),
+                 "'interactiveGroupColumn' must be of class 'character'")
+    args$interactiveGroupColumn <- c("col1", "col2")
+    expect_error(do.call(.checkArgumentsPDTMTptm, args),
+                 "'interactiveGroupColumn' must have length 1")
 
     ## seed
     args <- args0