+++ tests/testthat/test-runPDTMTAnalysis.R
@@ -28,6 +28,11 @@ test_that("runPDTMTAnalysis works", {
                                             separator = ";"),
         proteinIdCol = function(df) getFirstId(df, colName = "Accession",
                                                separator = ";"),
+        stringIdCol = function(df) combineIds(df, combineCols = c("Gene.Symbol", "Accession"),
+                                              combineWhen = "missing", makeUnique = FALSE),
+        modificationsCol = "Modifications.in.Master.Proteins",
+        excludeUnmodifiedPeptides = FALSE,
+        keepModifications = NULL,
         iColPattern = "^Abundance\\\\.F.+\\\\.Sample\\\\.",
         sampleAnnot = data.frame(
             sample = c("HIS4KO_S05", "HIS4KO_S06", "HIS4KO_S07", "HIS4KO_S08",
@@ -64,6 +69,8 @@ test_that("runPDTMTAnalysis works", {
         volcanoS0 = 0.1,
         volcanoFeaturesToLabel = "",
         addInteractiveVolcanos = FALSE,
+        interactiveDisplayColumns = NULL,
+        interactiveGroupColumn = NULL,
         complexFDRThr = 0.1,
         maxNbrComplexesToPlot = Inf,
         seed = 42,
@@ -251,6 +258,39 @@ test_that("runPDTMTAnalysis works", {
     args$proteinIdCol <- 1
     expect_error(do.call(runPDTMTAnalysis, args),
                  "'proteinIdCol' must be of class 'character'")
+
+    ## stringIdCol
+    args <- args0
+    args$stringIdCol <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'stringIdCol' must be of class 'character'")
+
+    ## modificationsCol
+    args <- args0
+    args$inputLevel <- "PeptideGroups"
+    args$modificationsCol <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'modificationsCol' must be of class 'character'")
+    args$modificationsCol <- c("Modifications", "Modifications.in.Master.Proteins")
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'modificationsCol' must have length 1")
+
+    ## excludeUnmodifiedPeptides
+    args <- args0
+    args$inputLevel <- "PeptideGroups"
+    args$excludeUnmodifiedPeptides <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'excludeUnmodifiedPeptides' must be of class 'logical'")
+    args$excludeUnmodifiedPeptides <- c(TRUE, FALSE)
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'excludeUnmodifiedPeptides' must have length 1")
+
+    ## keepModifications
+    args <- args0
+    args$inputLevel <- "PeptideGroups"
+    args$keepModifications <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'keepModifications' must be of class 'character'")
 
     ## iColPattern
     args <- args0
@@ -575,6 +615,21 @@ test_that("runPDTMTAnalysis works", {
     args$addInteractiveVolcanos <- c(TRUE, FALSE)
     expect_error(do.call(runPDTMTAnalysis, args),
                  "'addInteractiveVolcanos' must have length 1")
+
+    # interactiveDisplayColumns
+    args <- args0
+    args$interactiveDisplayColumns <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'interactiveDisplayColumns' must be of class 'character'")
+
+    # interactiveGroupColumn
+    args <- args0
+    args$interactiveGroupColumn <- 1
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'interactiveGroupColumn' must be of class 'character'")
+    args$interactiveGroupColumn <- c("col1", "col2")
+    expect_error(do.call(runPDTMTAnalysis, args),
+                 "'interactiveGroupColumn' must have length 1")
 
     ## complexFDRThr
     args <- args0